+++ src/Phan/Language/EmptyUnionType.php
@@ -1242,6 +1242,18 @@ final class EmptyUnionType extends UnionType
     }
 
     /** @override */
+    public function hasMixedTypeStrict(): bool
+    {
+        return false;
+    }
+
+    /** @override */
+    public function hasMixedOrNonEmptyMixedType(): bool
+    {
+        return false;
+    }
+
+    /** @override */
     public function withFlattenedArrayShapeTypeInstances(): UnionType
     {
         return $this;