+++ src/Phan/Language/Type/NonEmptyMixedType.php
@@ -125,12 +125,6 @@ final class NonEmptyMixedType extends MixedType
 
     public function withIsNullable(bool $is_nullable): Type
     {
-        if ($is_nullable) {
-            if ($this->is_nullable) {
-                return $this;
-            }
-            return static::instance(false);
-        }
-        return NonNullMixedType::instance(false);
+        return $is_nullable === $this->is_nullable ? $this : self::instance($is_nullable);
     }
 }